Add files via upload
Quiz Application
Here you can provide the users with Quiz on various general knowledge
topics. These practices will ensure that you’re able to set the layouts properly
and slowly increase your pace of learning the Android application
development. In this you’ll learn to use various Layout components at the
same time understanding them better.



In the app, questions would be asked and answers would be shown as
multiple choices. The user selects the answer and gets shown on the screen
if the answers are correct. In the end the final marks would be shown to the
users.